Create Translation.py
This is just practice to become more familiar with python programming, essentially the user can create a text file that's written in English and the program will be able to return that file translated to all world languages. (languages.txt is a hard coded file. with the help of googles translator I simply needed the language code in order to translate properly)